Fixed bug in validateTimes function
The bug caused only the first passing time to be validated.
